Modify Folder Structure for Ease of Use
