--- backend/.env.sample
@@ -1,3 +1,0 @@
-SERVER_PORT=YOUR_SERVER_PORT
-DATABASE_URL=YOUR_MONGODB_URL
-YT_TOKEN=YOUR_YOUTUBE_TOKEN
