--- backend/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "sync-wave-backend",
-  "version": "0.1.0",
-  "private": true,
-  "main": "server.js",
-  "scripts": {
-    "start": "node server",
-    "dev": "nodemon server"
-  },
-  "dependencies": {
-    "axios": "^0.21.0",
-    "express": "^4.17.1",
-    "socket.io": "^2.3.0",
-    "uuid": "^8.3.1"
-  },
-  "devDependencies": {
-    "dotenv": "^8.2.0",
-    "nodemon": "^2.0.6"
-  }
-}
