+++ package.json
@@ -1,17 +1,24 @@
 {
-  "name": "sync-wave-frontend",
+  "name": "sync-wave",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@material-ui/core": "^4.11.0",
     "@testing-library/jest-dom": "^5.11.5",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
+    "axios": "^0.21.0",
+    "express": "^4.17.1",
+    "nodemon": "^2.0.6",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-scripts": "4.0.0",
-    "web-vitals": "^0.2.4"
+    "uuid": "^8.3.1",
+    "web-vitals": "^0.2.4",
+    "ws": "^7.4.0"
   },
   "scripts": {
+    "server-dev": "nodemon server/server",
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
