+++ server/server.js
@@ -3,13 +3,6 @@ require('dotenv').config();
 const express = require('express');
 const app = express();
 const http = require('http').createServer(app);
-const io = require('socket.io').listen(http);
-
-io.on('connection', (socket) => {
-  console.log('Sup');
-});
-
-app.use(express.static('D:\\TheVoid\\GitHub\\mobile-schedule-app\\build'));
 
 const roomRouter = require('./routes/room');
 app.use('/room', roomRouter);
