+++ src/index.js
@@ -1,0 +1,14 @@
+import { ThemeProvider } from '@material-ui/core';
+import React from 'react';
+import ReactDOM from 'react-dom';
+import App from './App';
+import darkTheme from './theme';
+
+ReactDOM.render(
+  <React.StrictMode>
+    <ThemeProvider theme={darkTheme}>
+      <App />
+    </ThemeProvider>
+  </React.StrictMode>,
+  document.getElementById('root')
+);
