+++ src/theme/index.js
@@ -1,0 +1,9 @@
+const { createMuiTheme } = require('@material-ui/core/styles');
+
+const darkTheme = createMuiTheme({
+  palette: {
+    type: 'dark',
+  },
+});
+
+export default darkTheme;
